+++ src/ipm/ipx/crossover.cc
@@ -27,7 +27,7 @@ void Crossover::PushAll(Basis* basis, Vector& x, Vector& y, Vector& z,
 
     // Run dual push phase.
     std::vector<Int> dual_superbasics;
-    for (Int p = 0; p < perm.size(); p++) {
+    for (size_t p = 0; p < perm.size(); p++) {
         Int j = perm[p];
         if (basis->IsBasic(j) && z[j] != 0.0)
             dual_superbasics.push_back(j);
@@ -111,7 +111,7 @@ void Crossover::PushPrimal(Basis* basis, Vector& x,
 
     control_.ResetPrintInterval();
     Int next = 0;
-    while (next < variables.size()) {
+    while (next < (Int)variables.size()) {
         if ((info->errflag = control_.InterruptCheck()) != 0)
             break;
 
@@ -248,7 +248,7 @@ void Crossover::PushDual(Basis* basis, Vector& y, Vector& z,
 
     control_.ResetPrintInterval();
     Int next = 0;
-    while (next < variables.size()) {
+    while (next < (Int)variables.size()) {
         if ((info->errflag = control_.InterruptCheck()) != 0)
             break;
 
@@ -344,7 +344,7 @@ void Crossover::PushDual(Basis* basis, Vector& y, Vector& z,
     const Vector& ub = model.ub();
 
     std::vector<int> sign_restrict(n+m);
-    for (Int j = 0; j < sign_restrict.size(); j++) {
+    for (size_t j = 0; j < sign_restrict.size(); j++) {
         if (x[j] != ub[j]) sign_restrict[j] |= 1;
         if (x[j] != lb[j]) sign_restrict[j] |= 2;
     }
